move logic to helper
